project/swelancer/requirements.txt
aiodebug
aiomonitor
ansible==11.1.0
async-lru
asyncio==3.4.3
asyncvnc
beautifulsoup4==4.12.3
blobfile
boostedblob
dill
docker
IPython
jupyter-client
libcst==1.5.1
loky
lxml==5.3.0
mitmproxy==11.0.2
moviepy==2.1.1
numpy==2.2.1
openai
pandas
playwright==1.47.0
pillow<11.0.0
pipx==1.7.1
psutil
pyarrow
pyotp==2.9.0
pyperclip==1.9.0
pytest-asyncio==0.25.0
pytest-playwright==0.6.2
pyyaml
pypdf==5.1.0
pusher==3.3.2
reportlab==4.2.5
requests==2.32.3
requests_toolbelt==1.0.0
service-identity
structlog
tenacity
vncdotool
